refactor (components): `Header` renamed file to `index`
